chore(space-wizard): refactor more widgets into independent components
Refactored the single and multi select dynamic form widgets into independent components.
