+++ src/app/space-wizard/components/app-generator-multiple-selection-searchable-list/app-generator-multiple-selection-searchable-list.component.scss
@@ -1,0 +1,96 @@
+$icon-path:'../../../../assets/icon/';
+@import '../../../../assets/stylesheets/base';
+@import '../../../../assets/stylesheets/_icons';
+@import '../../../../assets/stylesheets/shared/_layout';
+
+:host {
+  border:solid 1px #bbb;
+  overflow:hidden;
+
+@keyframes select-item {
+  from {
+    background-color: #bbb;
+  }
+  to {
+    background-color: #efefef;
+  }
+}
+
+.hidden {
+  visibility: hidden;
+}
+
+.selected {
+  background-color: #efefef;
+  animation-name: select-item;
+  animation-duration: 1s;
+}
+
+
+  .btn-left-multi-select-button {
+  //min-width: unset;
+  width:25px;
+  border-left-style: none;
+  border-top-style: none;
+  border-bottom: solid 3px #efefef;
+}
+
+.btn-right-multi-select-button {
+  //min-width: unset;
+  width:25px;
+  border-right-style: none;
+  border-top-style: none;
+  border-bottom: solid 3px #efefef;
+}
+
+.form-control-filter-input {
+  height: 2.8rem;
+  border: none;
+  border-bottom: solid 1px #ddd;
+  border-left: solid 1px #ddd;
+  border-right: solid 1px #ddd;
+  //margin-right: -2px;
+  //box-shadow: inset 0 2px 2px rgba(0, 0, 0, 0.075);
+}
+
+.included-list {
+  overflow: auto;
+  max-height: 30vh;
+  border-top: solid 1px #ddd;
+  border-bottom: solid 1px #ddd;
+  display:flex;
+  flex-direction:row;
+  flex-wrap:wrap;
+  justify-content:start;
+}
+
+.excluded-list {
+  overflow: auto;
+  max-height: 30vh;
+  box-shadow: inset 0 2px 2px rgba(0, 0, 0, 0.075);
+}
+
+.unselectedOption {
+  width: 100%;
+  border-bottom: solid 0 #ddd;
+  margin-bottom: 0;
+  font-weight: normal;
+}
+
+.selectedOption {
+  width: 100%;
+  border-bottom: solid 0 #ddd;
+  margin-bottom: 0;
+  font-weight: bold;
+
+}
+
+.input-group-addon{
+  width:25px;
+  border-top-width:0;
+  border-left-width:0;
+  border-right-width:0;
+
+}
+
+}
