+++ src/app/space-wizard/components/app-generator-single-selection-dropdown/app-generator-single-selection-dropdown.component.scss
@@ -1,0 +1,74 @@
+$icon-path:'../../../../assets/icon/';
+@import '../../../../assets/stylesheets/base';
+@import '../../../../assets/stylesheets/_icons';
+@import '../../../../assets/stylesheets/shared/_layout';
+
+:host {
+
+.combobox-container
+{
+  input[type='text']{
+    &.combobox{
+      background-color:$color-pf-white;
+      color:inherit;
+    }
+  }
+}
+.dropdown-menu
+{
+  display:block;
+  padding:0;
+  margin:0;
+  border-width:0;
+  overflow:auto;
+  height:0;
+  top:26px;
+  &.open{
+     border-width:1px;
+     height:30vh;
+  }
+  .dropdown-menu-item
+  {
+     padding:5px;
+     border-top:solid 1px transparent;
+     border-bottom:solid 1px transparent;
+     padding-bottom:1.0rem;
+
+
+    .item-name{
+       font-weight: 600;
+    }
+    .item-description{
+
+    }
+    &:hover{
+      background-color: $color-pf-blue-100;
+      border-top:solid 1px $color-pf-blue-200;
+      border-bottom:solid 1px $color-pf-blue-200;
+      .item-name{
+        color: $color-pf-black;
+      }
+      .item-description{
+        color: $color-pf-black;
+
+      }
+
+    }
+    &.selected
+    {
+      background-color: $color-pf-blue-300 !important;
+      .item-name{
+        color: $color-pf-white !important;
+      }
+      .item-description{
+        color: $color-pf-white !important;
+
+      }
+
+    }
+
+  }
+
+}
+
+}
