+++ src/app/space-wizard/components/forge-app-generator/forge-app-generator.component.ts
@@ -90,99 +90,6 @@ export class ForgeAppGeneratorComponent implements OnInit, OnDestroy, OnChanges 
     }
   }
 
-  allOptionsSelected(field: IField): boolean {
-    return !field.display.choices.find((i) => i.selected === false);
-  }
-
-  selectOption(field: IField, choice: IFieldChoice) {
-    choice.selected = true;
-    this.updateFieldValue(field);
-  }
-  toggleOption(field: IField, choice: IFieldChoice) {
-    choice.selected = !choice.selected;
-    this.updateFieldValue(field);
-  }
-
-  deselectOption(field: IField, choice: IFieldChoice) {
-    choice.selected = false;
-    this.updateFieldValue(field);
-  }
-
-  updateFieldValue(field: IField): IField {
-    if ( !field ) {
-      return null;
-    }
-    switch (field.display.inputType) {
-      case FieldWidgetClassificationOptions.MultipleSelection:
-      {
-        if ( field.display.hasChoices ) {
-          field.value = field.display.choices
-          .filter((o) => o.selected)
-          .map((o) => o.id);
-        } else {
-          field.value = [];
-        }
-        this.forgeClient.validate();
-        break;
-      }
-      default: {
-        break;
-      }
-    }
-    return field;
-  }
-
-  deselectAllOptions(field: IField) {
-    field.display.choices.forEach((o) => {
-      o.selected = false;
-    });
-  }
-
-  filterList(field: IField, filter: string) {
-    // TODO: better validation of bad chars
-    filter = filter.replace('*', '');
-    filter = filter.replace('?', '');
-    filter = filter.replace('/', '');
-    filter = filter.replace('\\', '');
-    let r = new RegExp(filter || '', 'ig');
-    field.display.choices.filter((o) => {
-      // set everything to not visible, except for selected
-      o.visible = false;
-      if (o.selected === true) {
-        o.visible = true;
-      }
-      // then match the input string
-      return ((o.id.match(r)) || []).length > 0;
-    })
-    .forEach(o => {
-      o.visible = true;
-    });
-
-  }
-
-  selectAllOptions(field: IField) {
-    field.display.choices.forEach((o) => {
-      o.selected = true;
-    });
-  }
-
-  toggleSelectAll(field: IField) {
-    if ( !field ) {
-      return;
-    }
-    // at least one not selected, then select all , else deselect all
-    let item = field.display.choices.find((i) => i.selected === false);
-    if ( item ) {
-      for ( let o of field.display.choices ) {
-        o.selected = true;
-      }
-    } else {
-      for ( let o of field.display.choices ) {
-        o.selected = false;
-      }
-    }
-    this.updateFieldValue(field);
-  }
 
   trackByFn(index: any, item: any) {
    return index;
@@ -191,8 +98,7 @@ export class ForgeAppGeneratorComponent implements OnInit, OnDestroy, OnChanges 
   private onWorkflowPropertyChanged(change?: INotifyPropertyChanged<IWorkflow>) {
     if ( change ) {
       if ( change.currentValue !== change.previousValue ) {
-        this.log(`
-          The workflow property changed value ...`);
+        this.log(`The workflow property changed value ...`);
         let current: IWorkflow = change.currentValue;
         this.forgeClient.workflow = current;
         this.subscribeToIncomingWorkflowTransitions(current);
